Rename consensus types module from types.py to consensus_types.py
